refactor(demo): read the clock once and move the seed builders to the domain
Four independent clock reads decided the budget window, the pointage boundary,
the lissage window and the goal horizons; a seed straddling midnight disagreed
with itself. The builders that read it are pure, so they leave the use case for
the domain, where they now take the instant instead of asking for it.
